+++ MetalSplatter/Resources/MultiStageRenderPath.metal
@@ -35,8 +35,8 @@ inline FragmentIn multiStageSplatVertex(uint vertexID,
                                         constant Splat* splatArray,
                                         constant UniformsArray &uniformsArray,
                                         constant int32_t* sortedIndices,
-                                        const device uint *editStates,
-                                        const device uint *transformIndices,
+                                        const device uchar *editStates,
+                                        const device ushort *transformIndices,
                                         const device float4x4 *transformPalette) {
     Uniforms uniforms = uniformsArray.uniforms[min(int(amplificationID), kMaxViewCount - 1)];
 
@@ -99,8 +99,8 @@ vertex FragmentIn multiStageSplatVertexShaderEditing(uint vertexID [[vertex_id]]
                                                      constant Splat* splatArray [[ buffer(BufferIndexSplat) ]],
                                                      constant UniformsArray & uniformsArray [[ buffer(BufferIndexUniforms) ]],
                                                      constant int32_t* sortedIndices [[ buffer(BufferIndexSortedIndices) ]],
-                                                     const device uint *editStates [[ buffer(BufferIndexEditState) ]],
-                                                     const device uint *transformIndices [[ buffer(BufferIndexTransformIndex) ]],
+                                                     const device uchar *editStates [[ buffer(BufferIndexEditState) ]],
+                                                     const device ushort *transformIndices [[ buffer(BufferIndexTransformIndex) ]],
                                                      const device float4x4 *transformPalette [[ buffer(BufferIndexTransformPalette) ]]) {
     return multiStageSplatVertex(vertexID,
                                  instanceID,
